requirements.txt
python-dotenv  #import dotenv
requests